+++ R/utils.R
@@ -1,16 +1,4 @@
 
-# Whether a package is loaded
-core_loaded <- function() {
-  search <- paste0("package:", core)
-  core[search %in% search()]
-}
-
-# Whether a package is unloaded
-core_unloaded <- function() {
-  search <- paste0("package:", core)
-  core[!search %in% search()]
-}
-
 # Grab a package version
 MomX_package_version <- function(x) {
   version <- as.character(unclass(utils::packageVersion(x))[[1]])
@@ -21,38 +9,8 @@ MomX_package_version <- function(x) {
   paste0(version, collapse = ".")
 }
 
-# Handles colored message with text_col
-msg <- function(..., startup = FALSE) {
-  message(text_col(...))
+
+is_attached <- function(x) {
+  paste0("package:", x) %in% search()
 }
 
-
-text_col <- function(x) {
-  # If RStudio not available, messages already printed in black
-  if (!rstudioapi::isAvailable()) {
-    return(x)
-  }
-
-  if (!rstudioapi::hasFun("getThemeInfo")) {
-    return(x)
-  }
-
-  theme <- rstudioapi::getThemeInfo()
-
-  if (isTRUE(theme$dark)) crayon::white(x) else crayon::black(x)
-
-}
-
-invert <- function(x) {
-  if (length(x) == 0) return()
-  stacked <- utils::stack(x)
-  tapply(as.character(stacked$ind), stacked$values, list)
-}
-
-
-style_grey <- function(level, ...) {
-  crayon::style(
-    paste0(...),
-    crayon::make_style(grDevices::grey(level), grey = TRUE)
-  )
-}