+++ NEWS.md
@@ -1,3 +1,3 @@
 # MomX 0.0.0.9000
 
-* Initial working release (Momocs only)
+* Added a `NEWS.md` file to track changes to the package.